% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_frame.R
\name{add_plot_frame}
\alias{add_plot_frame}
\title{Add a frame around a plot}
\usage{
add_plot_frame(sides = c(1:4), kula = "black")
}
\arguments{
\item{sides}{vector: Which sides should the frame be added to? Bottom (1),
left (2), top (3), and / or right (4). Default is to plot all four.}

\item{kula}{What colour should the frame be?}
}
\description{
Add a frame around a plot
}
