% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{as_figure}
\alias{as_figure}
\title{Generate a figure object.}
\usage{
as_figure(ggplot_object, label, w, h, unit = c("mm", "cm", "in"))
}
\arguments{
\item{ggplot_object}{A ggplot/cowplot object}

\item{label}{The final file name}

\item{w}{The final width after saving/rendering}

\item{h}{The final height after saving/rendering}

\item{unit}{The unit of h and w}
}
\value{
An object of class 'figure'
}
\description{
Generate a figure object.
}
