% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_mdbib}
\alias{as_mdbib}
\alias{as_mdbib.data.frame}
\title{Convert an object to an mdbib class instance.}
\usage{
as_mdbib(x, ...)

\method{as_mdbib}{data.frame}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of \code{\link{mdbib}} class.
}
\description{
Converts an object (usually an expression)
to an mdbib instance.
}
