% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_mdexpr}
\alias{as_mdexpr}
\title{Convert an object to an mdexpr class instance.}
\usage{
as_mdexpr(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of mdexpr class (\code{\link{mdexpr}})
}
\description{
Converts an object (usually an expression)
to an mdexpr instance.
}
