% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_mdtable.matrix}
\alias{as_mdtable.matrix}
\title{Convert a matrix into an mdtable object.}
\usage{
\method{as_mdtable}{matrix}(x, label, ref_name, caption, ...)
}
\arguments{
\item{x}{a data frame or tibble.}

\item{label}{the final file name.}

\item{ref_name}{name of the reference to the table,
used mainly in markdown documents.}

\item{caption}{text to be presented in the table caption.}

\item{...}{extra arguments passed to \code{\link[base]{as.data.frame}}}
}
\value{
An object of class 'mdtable'.
}
\description{
Converts a matrix into a data frame and bundles it with the
R markdown reference, caption and file name for saving on the disc.
}
\details{
See: \code{\link{mdtable}} for details.
}
