% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_mdtable}
\alias{as_mdtable}
\alias{as_mdtable.data.frame}
\alias{as_mdtable.matrix}
\title{Convert a data frame into an mdtable object.}
\usage{
as_mdtable(x, ...)

\method{as_mdtable}{data.frame}(x, label, ref_name, caption, ...)

\method{as_mdtable}{matrix}(x, label, ref_name, caption, ...)
}
\arguments{
\item{x}{a data frame or tibble.}

\item{...}{extra arguments. In case of matrices, extra arguments passed to
\code{\link[base]{as.data.frame}}}

\item{label}{the final file name.}

\item{ref_name}{name of the reference to the table,
used mainly in markdown documents.}

\item{caption}{text to be presented in the table caption.}
}
\value{
An object of class \code{\link{mdtable}}.
}
\description{
Bundles a data frame with its R markdown reference and caption
and file name for saving on the disc.
}
\details{
Technically, a data frame with label, ref_name
and caption attributes.
The \code{\link{mdtable}} class inherits from data frame most of its methods.
ref_name: needs to be a valid markdown/bookdown reference name. By
default, spaces, slashes and underscores are silently turned into '-'.
}
