% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_tools.R
\name{build_chunk}
\alias{build_chunk}
\title{Create an R markdown code chunk.}
\usage{
build_chunk(
  quosure,
  ref_name = NULL,
  include = NULL,
  echo = NULL,
  warning = NULL,
  message = NULL
)
}
\arguments{
\item{quosure}{a quosure object.}

\item{ref_name}{reference name, skipped if NULL.}

\item{include}{include chunk option, skipped if NULL.}

\item{echo}{echo chunk option, skipped if NULL.}

\item{warning}{warning chunk option, skipped if NULL.}

\item{message}{message chunk option, skipped if NULL.}
}
\value{
a chunk text.
}
\description{
Builds a standard code chunk with the user-specified options
based on the quosure object provided.
}
