% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_tools.R
\name{build_inline}
\alias{build_inline}
\title{Create an inline R code text.}
\usage{
build_inline(quosure)
}
\arguments{
\item{quosure}{a quosure object.}
}
\value{
inline code text.
}
\description{
Creates a standard inline R code based on a quosure.
}
