% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_intervals.R
\name{calc_intervals}
\alias{calc_intervals}
\title{Calculate pretty intervals and include the boundaries exactly}
\usage{
calc_intervals(
  x1,
  x2,
  intMin = 5,
  intMax = 10,
  forceZero = NULL,
  preferError = FALSE
)
}
\arguments{
\item{x1}{The first value. If less than x2, an ascending vector is
returned; if greater than x2, a descending vector is returned.}

\item{x2}{The second value.}

\item{intMin}{The minimum number of numbers in the vector.}

\item{intMax}{The maximum number of numbers in the vector. The maximum is
100, so that it is not left in an eternal searching loop.}

\item{forceZero}{Does zero have to be included in the vectors? By default,
if x1 and x2 fall either side of 0, this will be TRUE, and if both x1 and
x2 have the same sign, it will be FALSE. Can be overwritten.}

\item{preferError}{If a suitable interval cannot be found with the existing
'intMax', is it preferable to throw an error (TRUE), or ignore the
'intMax' argument (FALSE; default)?}
}
\description{
This function is my attempt at creating pretty intervals. It
  differs from `pretty()` because it forces the min and max values to be
  included in the returned vector. It still needs further testing.
}
