% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{collapse_and}
\alias{collapse_and}
\title{Collapse elements of a text vector with comma/and.}
\usage{
collapse_and(x, oxford = TRUE)
}
\arguments{
\item{x}{a text vector or a vector coercible to a character.}

\item{oxford}{logical, should the 'and' be placed according to the Oxford
comma rule?}
}
\value{
a character string.
}
\description{
Collapses elements of a text vector with commas and inserts 'and' before
the last element.
}
