% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{components.figure}
\alias{components.figure}
\alias{plot.figure}
\alias{components.mdtable}
\alias{components.mdexpr}
\alias{components.mdlink}
\title{Extract components of a figure, mdtable, mdexpr or mdlink object.}
\usage{
\method{components}{figure}(
  object,
  what = c("plot", "label", "w", "h", "unit", "ref_name", "caption", "legend_text"),
  ...
)

\method{plot}{figure}(x, ...)

\method{components}{mdtable}(object, what = c("label", "ref_name", "caption"), ...)

\method{components}{mdexpr}(
  object,
  what = c("quosure", "expression", "result", "ref_name", "caption"),
  ...
)

\method{components}{mdlink}(object, what = c("ref_name", "URL"), ...)
}
\arguments{
\item{object}{an object.}

\item{what}{The requested feature.
'plot': the ggplot,
'label': the rendering/saving label,
'h': rendering height,
'w': the rendering width,
'unit': the rendering unit,
'ref_name': reference name used in R markdown,
'caption' object reference caption used in R markdown,
'legend_text' extended legend text used in R markdown.
'quosure': the quosure,
'expression': R expression,
'result': the evaluation result,
'URL': the link's URL.}

\item{...}{extra arguments, currently none.}

\item{x}{an object.}
}
\value{
The requested feature
}
\description{
Accesses specific components of an object specified by the
`what` argument or plots it.
}
