% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{convert.figure}
\alias{convert.figure}
\title{Convert plot units.}
\usage{
\method{convert}{figure}(x, to = c("mm", "cm", "in"))
}
\arguments{
\item{x}{A figure object}

\item{to}{the requested unit}
}
\value{
A figure object with the dimensions converted to the requested unit
}
\description{
Convert plot units.
}
