% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{convert}
\alias{convert}
\title{Convert plot units.}
\usage{
convert(x, ...)
}
\arguments{
\item{x}{An object of interest}

\item{...}{Arguments passed to specific methods}
}
\value{
The object with the value converted to the requested unit
}
\description{
Convert plot units.
}
