% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{extract.figure}
\alias{extract.figure}
\title{Extract features of a figure object.}
\usage{
\method{extract}{figure}(
  x,
  what = c("plot", "label", "w", "h", "unit", "ref_name", "caption", "legend"),
  ...
)
}
\arguments{
\item{x}{A figure object}

\item{what}{The requested feature.
'plot': the ggplot,
'label': the rendering/saving label,
'h': rendering height,
'w': the rendering width,
'unit': the rendering unit,
'ref_name': reference name used in R markdown,
'caption' object reference caption used in R markdown,
'legend' extended legend text used in R markdown.}

\item{...}{extra arguments, currently none.}
}
\value{
The requested feature
}
\description{
Extract features of a figure object.
}
