% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{extract.mdexpr}
\alias{extract.mdexpr}
\title{Extract features of an mdexpr object.}
\usage{
\method{extract}{mdexpr}(
  x,
  what = c("quosure", "expression", "result", "ref_name", "caption"),
  ...
)
}
\arguments{
\item{x}{An mdexpr object}

\item{what}{The requested feature.
'quosure': the quosure,
'expression': expression,
'result': the evaluation result,
'ref_name': reference name used in R markdown,
'caption' object reference caption used in R markdown,}

\item{...}{extra arguments, currently none.}
}
\value{
The requested feature
}
\description{
Extract features of an mdexpr object.
}
