% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{extract.mdlink}
\alias{extract.mdlink}
\title{Extract features of an mdlink object.}
\usage{
\method{extract}{mdlink}(x, what = c("ref_name", "URL"), ...)
}
\arguments{
\item{x}{an mdlink object.}

\item{what}{feature to be extracted:
'ref_name': reference name used in R markdown or
'URL': the link's URL.}

\item{...}{extra arguments, currently none.}
}
\description{
Extract features of an mdlink object.
}
