% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{extract.mdtable}
\alias{extract.mdtable}
\title{Extract features of an mdtable object.}
\usage{
\method{extract}{mdtable}(x, what = c("label", "ref_name", "caption"), ...)
}
\arguments{
\item{x}{An mdtable object}

\item{what}{The requested feature.
'label': the rendering/saving label,
'ref_name': reference name used in R markdown,
'caption' object reference caption used in R markdown,}

\item{...}{extra arguments, currently none.}
}
\value{
The requested feature
}
\description{
Extract features of an mdtable object.
}
