% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{figure}
\alias{figure}
\title{Generate a figure object.}
\usage{
figure(
  x,
  label,
  w,
  h,
  unit = c("mm", "cm", "in"),
  ref_name = NULL,
  caption = NULL,
  legend_text = NULL
)
}
\arguments{
\item{x}{a ggplot/cowplot object}

\item{label}{the final file name}

\item{w}{the final width after saving/rendering}

\item{h}{the final height after saving/rendering}

\item{unit}{the unit of h and w, mm be default}

\item{ref_name}{optional, name of the reference to the figure,
used mainly in markdown documents.}

\item{caption}{optional, text to be presented in the figure caption.}

\item{legend}{optional, text to be presented in the extended figure legend.}
}
\value{
An object of class 'figure'
}
\description{
Bundles a ggplot/cowplot object, file reference, the final
rendered graphic dimensions and markdown references in one figure object.
}
\details{
ref_name: needs to be a valid markdown/bookdown reference name. By
default, spaces, slashes and underscores are silently turned into '-'.
You may easily insert the read-to-use figure object code chunk into your
Rmarkdown document with the `insert()` method or reference it in the text
by calling `refer()`.
}
