% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{height}
\alias{height}
\title{Get object height.}
\usage{
height(x, ...)
}
\arguments{
\item{x}{An object of interest}

\item{...}{Arguments passed to specific methods}
}
\value{
The object's height
}
\description{
Get object height.
}
