% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{html_italic}
\alias{html_italic}
\alias{html_bold}
\alias{html_sup}
\alias{html_sub}
\title{Ready-to-use HTML expressions.}
\usage{
html_italic(x)

html_bold(x)

html_sup(x)

html_sub(x)
}
\arguments{
\item{x}{a character string or an object which can be converted to a
character string.}
}
\value{
a string
}
\description{
This function set generates HTML expressions which may be used in e.g. in
titles, labels and text of `ggplot` objects in combination with `ggtext`
package's `element_markdown`.
}
