% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{insert.mdexpr}
\alias{insert.mdexpr}
\title{Insert a code chunk into an R markdown file}
\usage{
\method{insert}{mdexpr}(
  object,
  file = NULL,
  format = c("chunk", "inline"),
  include = NULL,
  echo = NULL,
  warning = NULL,
  message = NULL,
  append = TRUE,
  ...
)
}
\arguments{
\item{object}{an mdexpr object (\code{\link{mdexpr}}).}

\item{file}{a file to which the chunks are written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console and copied into the clipboard.}

\item{format}{which format the table should be inserted in?
'chunk' (default): a multiline code chunk or 'inline' code piece.}

\item{include}{the include option of the chunk, skipped if NULL.}

\item{echo}{echo chunk option, skipped if NULL.}

\item{warning}{warning chunk option, skipped if NULL.}

\item{message}{message chunk option, skipped if NULL.}

\item{append}{logical, should the output file be appended?}

\item{...}{extra arguments, currently none.}
}
\description{
Inserts a code chunk based on an mdexpr object
(\code{\link{mdexpr}}) into an R markdown file or prints it into the
standard output.
}
\details{
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}. If the ref_name parameter is set, the code
chunk will be named after it, otherwise the name is absent.
For include, echo, message and warning chunk options, see:
https://rmarkdown.rstudio.com/lesson-3.html
}
