% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{insert.mdhtml}
\alias{insert.mdhtml}
\title{Insert an HTML element.}
\usage{
\method{insert}{mdhtml}(object, file = NULL, append = TRUE, ...)
}
\arguments{
\item{object}{an mdhtml object (\code{\link{mdhtml}}).}

\item{file}{a file to which the chunks are written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console and copied into the clipboard.}

\item{append}{logical, should the output file be appended?}

\item{...}{extra arguments, currently none.}
}
\description{
Inserts a HTML elemen based on an mdhtml object
(\code{\link{mdhtml}}) into an R markdown file or prints it into the
standard output.
}
\details{
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
}
