% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{insert.mdlink}
\alias{insert.mdlink}
\alias{refer.mdlink}
\title{Insert/refer a link.}
\usage{
\method{insert}{mdlink}(
  object,
  title = object$ref_name,
  html = FALSE,
  file = NULL,
  append = TRUE,
  ...
)

\method{refer}{mdlink}(
  object,
  title = object$ref_name,
  html = FALSE,
  file = NULL,
  append = TRUE,
  ...
)
}
\arguments{
\item{object}{an mdlink object (\code{\link{mdlink}}).}

\item{title}{a string to be used as a link title, defaults to the `ref_name`
of the mdlink object.}

\item{html}{logical, should the link be in a HTML format? Defaults to FALSE.}

\item{file}{a file to which the chunks are written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console and copied into the clipboard.}

\item{append}{logical, should the output file be appended?}

\item{...}{extra arguments, currently none.}
}
\description{
Inserts a link based on an mdlink object
(\code{\link{mdlink}}) into an R markdown file or prints it into the
standard output.
}
\details{
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
}
