% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{insert.mdtext}
\alias{insert.mdtext}
\alias{refer.mdtext}
\title{Insert/refer a text element.}
\usage{
\method{insert}{mdtext}(object, file = NULL, append = TRUE, ...)

\method{refer}{mdtext}(object, file = NULL, append = TRUE, ...)
}
\arguments{
\item{object}{an mdtext object (\code{\link{mdtext}}).}

\item{file}{a file to which the chunk will be written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console and copied into the clipboard.}

\item{append}{logical, should the output file be appended?}

\item{...}{extra arguments, currently none.}
}
\description{
Inserts a text element based on an mdtext object
(\code{\link{mdtext}}) into an R markdown file or prints it into the
standard output.
}
\details{
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
}
