% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{insert}
\alias{insert}
\title{Insert an object chunk into an R markdown file}
\usage{
insert(object, ...)
}
\arguments{
\item{object}{an object to be referenced.}

\item{...}{extra arguments passed to methods.}
}
\description{
Inserts a reference to the object as an R code chunk into
a R markdown file or prints it into the standard output.
}
