% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_tools.R
\name{insert_figure}
\alias{insert_figure}
\title{Insert a reference to figure objects into a Rmarkdown file.}
\usage{
insert_figure(
  ...,
  file = NULL,
  ref_names = NULL,
  captions = NULL,
  legend = TRUE,
  legend_text = "<<legend>>",
  html = FALSE,
  style_ref = "legend",
  append = FALSE
)
}
\arguments{
\item{...}{figure objects.}

\item{file}{a file to which the chunks are written. If the file exists
already, it will be appended or overwritten. If NULL, the text is printed
in the console.}

\item{ref_names}{names of the figure chunk in the R markdown output.
If NULL, default names ('figure1', 'figure2' and so on) are used.}

\item{captions}{figure captions. If NULL, default captions
('caption for figure 1' and so on) are used.}

\item{legend}{logical, should a text with the figure reference in bold be
included below the figure chunk?}

\item{legend_text}{a character vector or list with detailed legend texts.}

\item{html}{logical: should the legend be HTML styled?}

\item{style_ref}{name of the CSS style of the legend text, valid only for
the HTML output.}

\item{append}{logical, should the output file be appended?}
}
\value{
invisibly retuns a ready-to-use figure chunk text.
}
\description{
Builds a character with the R code chunks for the given figure objects.
}
