% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_figure}
\alias{is_figure}
\alias{is_mdtable}
\alias{is_mdexpr}
\alias{is_mdlink}
\alias{is_mdhtml}
\alias{is_mdtext}
\alias{is_mdbib}
\title{Check for class inheritance}
\usage{
is_figure(x)

is_mdtable(x)

is_mdexpr(x)

is_mdlink(x)

is_mdhtml(x)

is_mdtext(x)

is_mdbib(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical, TRUE if the object belongs to the class of interest.
}
\description{
Check if an object inherits from a class of interest.
}
