% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_mdbib}
\alias{is_mdbib}
\title{Check for a mdbib object.}
\usage{
is_mdbib(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, TRUE if the 'mdbib' class object provided.
}
\description{
Check for a mdbib object.
}
