% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_mdexpr}
\alias{is_mdexpr}
\title{Check for a mdexpr object.}
\usage{
is_mdexpr(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, TRUE if the 'figure' class object
}
\description{
Check for a mdexpr object.
}
