% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_mdhtml}
\alias{is_mdhtml}
\title{Check for a mdhtml object.}
\usage{
is_mdhtml(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, TRUE if the 'mdhtml' class object provided.
}
\description{
Check for a mdhtml object.
}
