% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_mdlink}
\alias{is_mdlink}
\title{Check for a mdlink object.}
\usage{
is_mdlink(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, TRUE if the 'mdlink' class object provided.
}
\description{
Check for a mdlink object.
}
