% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_mdtable}
\alias{is_mdtable}
\title{Check for a mdtable object.}
\usage{
is_mdtable(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, TRUE if the 'figure' class object
}
\description{
Check for a mdtable object.
}
