% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_mdtext}
\alias{is_mdtext}
\title{Check for a mdtext object.}
\usage{
is_mdtext(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, TRUE if the 'mdtext' class object provided.
}
\description{
Check for a mdtext object.
}
