% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mdbib}
\alias{mdbib}
\title{Create an mdbib object storing document's bibliography}
\usage{
mdbib(x, ...)
}
\arguments{
\item{x}{a data frame. It has to contain the 'BIBTEXKEY' variable.}

\item{...}{additional arguments, currently none defined.}
}
\value{
an object of the 'mdbib' class.
}
\description{
Creates an mdbib object storing bibliography information
(preferably derived from a BibTex file). Takes a data frame or a similar
structure. The variable named 'BIBTEXKEY' is required.
}
\details{
The object is built on the top of a data frame and inherits
multiple functions/methods from the 'data_frame' class.
}
