% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mdexpr}
\alias{mdexpr}
\title{Generate an mdexpr object.}
\usage{
mdexpr(x, ref_name = NULL, caption = NULL, ...)
}
\arguments{
\item{x}{an expression.}

\item{ref_name}{name of the reference to the code block,
used mainly in markdown documents (waiting for a bookdown implementation).}

\item{caption}{text to be presented in the table caption, currently
no implementation in bookdown
(https://github.com/rstudio/bookdown/issues/238)}

\item{...}{extra arguments, currently none.}
}
\value{
an object of mdexpr class.
}
\description{
Generates an mdexpression object which bundles an R expression
with its evaluation result and enables seamless insertion
into an R markdown document.
}
\details{
Throws any errors and warnings encountered during the evaluation.
In addition checks if the result is NULL, NA or Inf and raises a warning
if this is the case.

ref_name: needs to be a valid markdown/bookdown reference name. By
default, spaces, slashes and underscores are silently turned into '-'.
}
