% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mdhtml}
\alias{mdhtml}
\title{Generate an mdhtml object.}
\usage{
mdhtml(x)
}
\arguments{
\item{x}{a string to be stored as a HTML or XTML tag, see the details!}
}
\description{
Generates an mdhtml object which stores a HTML or XTML tag
}
\details{
Particularly useful for custom HTML element used multiple times
in an Rmarkdown document. You may easily insert them (e.g. via copy and paste)
by calling the `insert()` or `refer()` method. Technically,
it requires a string starting with an `<tag>` or `<tag />` and ending with
an `</tag>` or `<tag />`.
}
