% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mdlink}
\alias{mdlink}
\title{Generate an mdlink object.}
\usage{
mdlink(x, ref_name)
}
\arguments{
\item{x}{a URL}

\item{ref_name}{a link title}
}
\description{
Generates an mdlink object which holds together
the URL and tile of a link.
}
\details{
Particularly useful for link used multiple times
in an Rmarkdown document. You may easily insert them (e.g. via copy and paste)
by calling the `insert()` method.
}
