% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mdtable}
\alias{mdtable}
\title{Generate an mdtable object.}
\usage{
mdtable(x, label, ref_name, caption)
}
\arguments{
\item{x}{a data frame or tibble.}

\item{label}{the final file name.}

\item{ref_name}{name of the reference to the table,
used mainly in markdown documents.}

\item{caption}{text to be presented in the table caption.}
}
\value{
An object of class 'mdtable'
}
\description{
Bundles a data frame with its R markdown reference and caption
and file name for saving on the disc.
}
\details{
Technically, a data frame with label, ref_name
and caption attributes.
The mdtable inherits from data frame most of its methods.
ref_name: needs to be a valid markdown/bookdown reference name. By
default, spaces, slashes and underscores are silently turned into '-'.
}
