% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mdtext}
\alias{mdtext}
\title{Create an mdtext container.}
\usage{
mdtext(x)
}
\arguments{
\item{x}{a string to be stored as mdtext object.}
}
\description{
Creates a container for custom text to be (re-) used in the
Rmarkdown document.
}
\details{
Storing some fixed text parts used multiple times in the Rmarkdown
document (e.g. parts of figure legends) as mdtext provides a smarter
alternative to the tarditional 'copy-paste' approach.
You may insert the text chunk in your document by calling the
`insert()` or `refer()` method.
}
