% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{number_figures}
\alias{number_figures}
\title{Set consecutive figure numbers in a list of `figure` objects.}
\usage{
number_figures(figure_lst, prefix = "figure_")
}
\arguments{
\item{figure_lst}{a list of \code{\link{figure}} objects.}

\item{prefix}{a character string that precedes the figure number.}
}
\value{
a list of \code{\link{figure}} objects.
}
\description{
Appends labels of `figure` objects in a list with consecutive numbers (e.g.
'figure_1_...', 'figure_2_...').
}
