% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{pickle.figure}
\alias{pickle.figure}
\title{Save a figure object on the disc.}
\usage{
\method{pickle}{figure}(object, path = ".", format = "pdf", ...)
}
\arguments{
\item{object}{a figure object.}

\item{path}{The target path. The destination folder will not be created}

\item{format}{the desired file format.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{ggsave}}.}
}
\value{
None, called for side effects
}
\description{
Saves a figure object on the disc.
}
