% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{pickle.mdtable}
\alias{pickle.mdtable}
\title{Save an mdtable object on the disc.}
\usage{
\method{pickle}{mdtable}(object, folder = ".", format = "tsv", delim = "\\t", ...)
}
\arguments{
\item{object}{an mdtable object.}

\item{folder}{The target folder. The destination folder will not be created.}

\item{format}{the desired file format.}

\item{delim}{delimiter for clumns separation, tabulation by default.}

\item{...}{extra arguments passed to \code{\link[readr]{write_delim}}.}
}
\value{
None, called for side effects
}
\description{
Saves an mdtable object on as a text file on the disc.
}
