% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pickle}
\alias{pickle}
\title{Save and object on the disc.}
\usage{
pickle(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Saves an object on the disc.
}
