% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot.figure}
\alias{plot.figure}
\title{Plot a figure object.}
\usage{
\method{plot}{figure}(figure_object)
}
\arguments{
\item{figure_object}{A figure object}
}
\value{
The figure's ggplot object
}
\description{
Plot a figure object.
}
