% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Scatterplot of x and y values}
\usage{
plot_points(
  x,
  y,
  borderKula = "#004488FF",
  pointKula = "white",
  cex = 1.75,
  pch = 21,
  lwd = 3,
  xpd = TRUE,
  ...
)
}
\arguments{
\item{x}{The x data to plot.}

\item{y}{The y data to plot.}

\item{borderKula}{What colour should the boundary of the marker be?}

\item{pointKula}{What colour should the centre of the marker be?}

\item{cex}{What size should the marker be?}

\item{pch}{What style should the marker be?}

\item{lwd}{How thick should the marker boundary be?}

\item{xpd}{Can the markers be plotted off the chart?}
}
\description{
This is the figuR equivalent of `plot(x, y)` to create a
  simple scatterplot.
}
