% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_plot_matrix.R
\name{pre_plot_matrix}
\alias{pre_plot_matrix}
\title{Prepares to image-plot a matrix}
\usage{
pre_plot_matrix(x, ...)
}
\arguments{
\item{x}{matrix: The data that will be added afterwards.}

\item{...}{Any arguments (except 'xLimits' and 'yLimits') that can be added
to [pre_plot()].}
}
\description{
When using [pre_plot], it is not possible to subsequently plot
  a matrix using  `image(x, add = TRUE)` because the limits don't line up.
  However, it appears (at least so far) that the extension to each side is
  predictable, and depends on the number of rows and cols. This function
  accounts for this, and thus allows a matrix to be image plotted onto a
  customisable frame from [pre_plot()].

  This function is mainly intended to be used by [plot_matrix()].
}
