% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_plot_old.R
\name{pre_plot_old}
\alias{pre_plot_old}
\title{Fine-combed control for plots}
\usage{
pre_plot_old(
  xLimits,
  yLimits,
  main = "",
  mainOffset = 0.25,
  mainCex = 1.1,
  xTickSeq = NULL,
  xTickCount = NULL,
  xTickFirst = NULL,
  xTickEvery = NULL,
  xAlignMidPoints = FALSE,
  xTickLength = 0.2,
  xTickKula = "#1A1A1AFF",
  xAxisSide = 1,
  xLabels = NULL,
  xLabelEvery = NULL,
  xLabelFirst = NULL,
  xLabelOffset = NULL,
  xLabelCex = 0.92,
  xLabelSrt = 0,
  xLabelKula = "#4D4D4DFF",
  xAxisLwd = 1,
  xAxisType = 1,
  xAxisKula = "#1A1A1AFF",
  xName = NULL,
  xNameAxis = NULL,
  xNameKula = "#1A1A1AFF",
  xNameCex = 1,
  xNameOffset = NULL,
  xNameSrt = NULL,
  xGridEvery = NULL,
  xGridFirst = NULL,
  xGridLwd = 1,
  xGridType = 1,
  xGridKula = "#E6E6E6AA",
  yTickSeq = NULL,
  yTickCount = NULL,
  yTickFirst = NULL,
  yTickEvery = NULL,
  yAlignMidPoints = FALSE,
  yTickLength = 0.2,
  yTickKula = "#1A1A1AFF",
  yAxisSide = 2,
  yLabels = NULL,
  yLabelEvery = NULL,
  yLabelFirst = NULL,
  yLabelOffset = NULL,
  yLabelCex = 0.92,
  yLabelSrt = NULL,
  yLabelKula = "#4D4D4DFF",
  yAxisLwd = 1,
  yAxisType = 1,
  yAxisKula = "#1A1A1AFF",
  yName = NULL,
  yNameAxis = NULL,
  yNameKula = "#1A1A1AFF",
  yNameCex = 1,
  yNameOffset = NULL,
  yNameSrt = NULL,
  yGridEvery = NULL,
  yGridFirst = NULL,
  yGridLwd = 1,
  yGridType = 1,
  yGridKula = "#E6E6E6AA",
  addOrigin = TRUE,
  originLwd = 1,
  originType = 1,
  originKula = "#B3B3B388",
  add121 = FALSE,
  annotationLocation = c(0, 0),
  annotationText = "",
  annotationCex = 0.9,
  mar = c(2.5, 3.5, 2.5, 1.5)
)
}
\description{
Easily customisable plot areas
}
