% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{print.figure}
\alias{print.figure}
\alias{print.mdexpr}
\alias{print.mdlink}
\title{Print a figure, mdexpr or mdlink object}
\usage{
\method{print}{figure}(x, ...)

\method{print}{mdexpr}(x, ...)

\method{print}{mdlink}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{extra arguments, currently none.}
}
\value{
The figure's label and dimensions
}
\description{
Print a figure, mdexpr or mdlink object
}
