% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{print.mdexpr}
\alias{print.mdexpr}
\title{Print as mdexpr object.}
\usage{
\method{print}{mdexpr}(x, ...)
}
\arguments{
\item{x}{an mdexpr object.}

\item{...}{extra arguments, currently none.}
}
\description{
Prints the expression and evaluation result of an mdexpr
object.
}
