% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{print.mdlink}
\alias{print.mdlink}
\title{Print an mdlink object.}
\usage{
\method{print}{mdlink}(x, ...)
}
\arguments{
\item{x}{an mdlink object.}

\item{...}{extra arguments, currently none.}
}
\description{
displays an mdlink object.
}
