% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{read_bib}
\alias{read_bib}
\title{Read bibliography from a text file.}
\usage{
read_bib(file, ...)
}
\arguments{
\item{file}{character, path or URL to a .bib file.}

\item{...}{extra arguments passed to \code{\link[bib2df]{bib2df}}.}
}
\value{
an \code{\link{mdbib}} class object.
}
\description{
Reads a BibTex file from the disc and converts it to an
\code{\link{mdbib}} class object.
}
\details{
Technically, the function employs the \code{\link[bib2df]{bib2df}}
function.
}
