% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/oop.R
\name{refer}
\alias{refer}
\alias{refer.figure}
\alias{refer.mdtable}
\alias{refer.mdexpr}
\alias{refer.mdbib}
\alias{refer.mdtext}
\alias{refer.mdlink}
\alias{refer.mdhtml}
\title{Cite a figure, table, literature, text, URL link
or generate inline R/HTML code}
\usage{
refer(object, ...)

\method{refer}{figure}(object, file = NULL, append = TRUE, ...)

\method{refer}{mdtable}(object, file = NULL, append = TRUE, ...)

\method{refer}{mdexpr}(object, file = NULL, append = TRUE, ...)

\method{refer}{mdbib}(object, file = NULL, append = TRUE, ...)

\method{refer}{mdtext}(object, file = NULL, append = TRUE, ...)

\method{refer}{mdlink}(
  object,
  title = object$ref_name,
  html = FALSE,
  file = NULL,
  append = TRUE,
  ...
)

\method{refer}{mdhtml}(object, file = NULL, append = TRUE, ...)
}
\arguments{
\item{object}{an object to be referenced.}

\item{...}{extra arguments, currently none.}

\item{file}{a file to which the references or code are written.
If the file exists already, it will be appended or overwritten.
If NULL, the text is printed in the console and copied into the clipboard.}

\item{append}{logical, should the output file be appended?}

\item{title}{a string to be used as a link title, defaults to the `ref_name`
of the mdlink object.}

\item{html}{logical, should the link be in a HTML format? Defaults to FALSE.}
}
\description{
Generates an inline citation to a figure, mdtable object or
literature references, currently only in the bookdown style.
For mdbib objects, it inserts a citation to a all entries of
a bibliography entries (e.g. `[@citation1; @citation2; ...]`).
In combination with the \code{\link{reglook.mdbib}}
or the tidyverse's `filter()` method, it can be used to search
a bibliography object and paste the selected citations
into an Rmarkdown document.
For mdexpr, mdhtml and mdtext objects storing R, HTML code
and ordinary text, respectively, an inline text is generated.
For mdlinks a link in the markdown or html format is generated.
}
\details{
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
`refer()` is a S3 generic function.
}
