% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{refer.mdbib}
\alias{refer.mdbib}
\title{Reference to bibliography.}
\usage{
\method{refer}{mdbib}(object, file = NULL, append = TRUE, ...)
}
\arguments{
\item{object}{an mdbib object (\code{\link{mdbib}}).}

\item{file}{a file to which the chunk will be written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console and copied into the clipboard.}

\item{append}{logical, should the output file be appended?}

\item{...}{extra arguments, currently none.}
}
\description{
Inserts a citation to a all entries of a bibliography object
of the 'mdbib' class. In combination with the \code{\link{reglook.mdbib}}
or the tidyverse's `filter()` method, it can be used to search
a bibliography and paste the selected citations into an Rmarkdown document.
}
\details{
Creates Rmarkdown style citations: `[@citation1; @citation2; ...]`.

To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
}
