% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{refer.mdexpr}
\alias{refer.mdexpr}
\title{Insert inline code.}
\usage{
\method{refer}{mdexpr}(object, file = NULL, append = TRUE, ...)
}
\arguments{
\item{object}{an mdexpr object.}

\item{file}{a file to which the chunks are written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console.}

\item{append}{logical, should the output file be appended?}

\item{...}{extra arguments, currently none.}
}
\description{
Generates an inline code text for the provided mdexpr object.
}
\details{
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
}
