% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{refer.mdtable}
\alias{refer.mdtable}
\title{Cite an mdtable object.}
\usage{
\method{refer}{mdtable}(object, file = NULL, append = TRUE, ...)
}
\arguments{
\item{object}{an mdtable object.}

\item{file}{a file to which the chunks are written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console.}

\item{append}{logical, should the output file be appended?}

\item{...}{extra arguments, currently none.}
}
\description{
Generates an inline citation to an mdtable object, currently
only in the bookdown style.
}
\details{
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
}
