% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{refer}
\alias{refer}
\title{Insert an object citation/reference into an R markdown file}
\usage{
refer(object, ...)
}
\arguments{
\item{object}{am object to be referenced.}

\item{....}{extra arguments passed to methods.}
}
\description{
Inserts a reference to the object as an inline markdown
reference or prints it into the standard output.
}
