% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{reglook.mdbib}
\alias{reglook.mdbib}
\title{Search bibliography by a regular expression.}
\usage{
\method{reglook}{mdbib}(object, regex, keys = NULL, multiple = c("OR", "AND"))
}
\arguments{
\item{object}{an mdbib object (\code{\link{mdbib}}).}

\item{regex}{a regular expression.}

\item{keys}{columns of the mdbib object to be searched for the regex.
Defaults to NULL, which means that all columns will be looked up.}

\item{multiple}{ignored when a single search key is provided.
It describes how searching results from multiple mdbib columns will
be handled. 'OR' (default) specifies that the multiply column search results
will be merged by logical sum, for 'AND' they will be merged by logical
intersection.}

\item{...}{extra arguments, currently none.}
}
\value{
a mdbib object.
}
\description{
Searches the content of a bibliography object of the 'mdbib'
class. In combination with the \code{\link{refer.mdbib}} method,
it can be used to search a bibliography and paste the selected citations
into an Rmarkdown document.
}
\details{
Technically, the search is accomplished by the
\code{\link[stringi]{stri_detect}} function.
}
