% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{reglook}
\alias{reglook}
\title{Search an object by a regular expression.}
\usage{
reglook(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Searches the content of an object with a regular expression.
}
