% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{resize}
\alias{resize}
\title{Resize a figure object.}
\usage{
resize(
  figure_object,
  value,
  what = c("w", "h", "scale"),
  unit = extract(figure_object, "unit"),
  lock = TRUE
)
}
\arguments{
\item{figure_object}{A figure object}

\item{value}{The new dimension value}

\item{what}{The requested dimension to be changed or the scaling factor}

\item{unit}{The dimension unit of the final plot object}

\item{lock}{If TRUE, the aspect ratio is kept constant (scaling), ignored if 'what' is set to 'scale'}
}
\value{
A figure object with the modified dimensions and (optionally) unit
}
\description{
Resize a figure object.
}
