% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modif_functions.R
\name{save_figure}
\alias{save_figure}
\title{Save a figure object on the disc}
\usage{
save_figure(figure_object, path = ".", format = "pdf", ...)
}
\arguments{
\item{figure_object}{A figure object}

\item{path}{The target path. The destination folder will not be created}

\item{format}{the desired file format.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{ggsave}}.}
}
\value{
None, called for side effects
}
\description{
Save a figure object on the disc
}
