% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modif_functions.R
\name{save_mdtable}
\alias{save_mdtable}
\title{Save an md table object on the disc}
\usage{
save_mdtable(mdtable_object, folder = ".", format = "tsv", delim = "\\t", ...)
}
\arguments{
\item{mdtable_object}{A mdtable object.}

\item{folder}{The target folder. The destination folder will not be created.}

\item{format}{the desired file format.}

\item{delim}{delimiter for column separation, tabulation by default.}

\item{...}{extra arguments passed to \code{\link[readr]{write_delim}}.}
}
\value{
None, called for side effects
}
\description{
Save an md table object on the disc
}
