% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_regex}
\alias{search_regex}
\title{Search a vector with a regular expression.}
\usage{
search_regex(x, regex, ...)
}
\arguments{
\item{x}{a vector.}

\item{regex}{a regular expression.}

\item{...}{additional arguments passed to \code{\link[stringi]{stri_detect}}.}
}
\value{
a logical vector.
}
\description{
Searches a vector (character or list) with a regular expression.
}
