% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_tools.R
\name{set_widths}
\alias{set_widths}
\title{Set column widths in a flextable.}
\usage{
set_widths(x, widths, unit = c("cm", "mm", "in"))
}
\arguments{
\item{x}{a `flextable` object.}

\item{widths}{a numeric vector with widths of the columns, begining from
the first one.}

\item{unit}{unit of width, one of 'cm' (default), 'mm' or 'in'.}
}
\value{
a `flextable` class object.
}
\description{
Sets widths of columns in a `flextable` object.
}
