% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_lty.R
\name{show_lty}
\alias{show_lty}
\title{Plot each of the lty line styles}
\usage{
show_lty(lwd = 1, kula = "black")
}
\arguments{
\item{lwd}{numeric: How thick should the lines be?}

\item{kula}{What colours should the lines be?}
}
\description{
Who can ever remember the lty codes for the different line
  styles. This create a plot showing each of them in order.
}
