% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_pch.R
\name{show_pch}
\alias{show_pch}
\title{Plot each of the pch symbols}
\usage{
show_pch()
}
\description{
Who can ever remember the pch codes for the different point
  markers? This function creates a plot showing each of them in order.
}
