% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{stri_capitalize_first}
\alias{stri_capitalize_first}
\alias{stri_decapitalize_first}
\title{Capitalize and de-capitalize the first character of a string.}
\usage{
stri_capitalize_first(x)

stri_decapitalize_first(x)
}
\arguments{
\item{x}{a text vector or a vector coercible to a character.}
}
\value{
a character vector.
}
\description{
The functions capitalize or de-capitalize the first character of a string.
}
