% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tag2subtitle}
\alias{tag2subtitle}
\title{Move plot tag to subtitle.}
\usage{
tag2subtitle(
  x,
  line_replacement = ", ",
  tolower = FALSE,
  replace = TRUE,
  sep = "\\n",
  before = FALSE
)
}
\arguments{
\item{x}{a `ggplot` object.}

\item{line_replacement}{a string to replace new lines in the tag's body.}

\item{tolower}{logical, should the tag be decapitalized prior to moving
to the subtitle?}

\item{replace}{logical, should the existing subtitle be replaced?}

\item{sep}{separator for pasting the existing subtitle and the tag. Ignored
if `replace = TRUE`,}

\item{before}{logical, should the tag be pasted before the existing subtitle?
Ignored if `replace = TRUE`.}
}
\description{
Moves plot tag to subtitle. Any new line marks at the beginning of
the tag will be removed. New line marks in the body of the tag will be
replaced by the string provided as `line_replacement`.
If `replace = TRUE`, the existing subtitle will be replaced by the tag.
Otherwise, it will be pasted with the existing one with a separator defined
by the `sep` argument.#'
}
