% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{width}
\alias{width}
\title{Get object width.}
\usage{
width(x, ...)
}
\arguments{
\item{x}{An object of interest}

\item{...}{Arguments passed to specific methods}
}
\value{
The object's width
}
\description{
Get object width.
}
