% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_figure}
\alias{as_figure}
\title{Convert and object to a figure class instance.}
\usage{
as_figure(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of figure class (\code{\link{figure}})
}
\description{
Converts an object (in most cases a graphic object)
to a figure object.
}
