% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_mdbib}
\alias{as_mdbib}
\title{Convert an object to an mdbib class instance.}
\usage{
as_mdbib(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of mdbib class (\code{\link{mdbib}})
}
\description{
Converts an object (usually an expression)
to an mdbib instance.
}
