% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_mdtable}
\alias{as_mdtable}
\title{Convert and object to an mdtable class instance.}
\usage{
as_mdtable(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
an object of mdtable class (\code{\link{mdtable}})
}
\description{
Converts an object (table or matrix in most cases) to an
mdtable instance.
}
