% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_tools.R
\name{build_fig_back}
\alias{build_fig_back}
\title{Create a R markdown backbone for a figure call.}
\usage{
build_fig_back(
  figure_call,
  ref_name = "figure",
  caption = "figure",
  figure_w = 5,
  figure_h = 7,
  legend = TRUE
)
}
\arguments{
\item{figure_call}{call to a figure object.}

\item{ref_name}{name of the figure chunk in the R markdown output.}

\item{caption}{figure caption.}

\item{figure_w}{figure width in inches.}

\item{figure_h}{figure height in inches.}

\item{legend}{logical, should a text with the figure reference in bold be included below
the figure chunk?}
}
\value{
a string with the figure object R markdown chunk.
}
\description{
Builds a character with the R code chunk for the given figure.
}
