% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_plot_limits.R
\name{calc_plot_limits}
\alias{calc_plot_limits}
\title{Calculate axis limits automatically}
\usage{
calc_plot_limits(x, y)
}
\arguments{
\item{x}{The x values that will be plotted.}

\item{y}{The y values that will be plotted.}
}
\description{
This function is used in `plot_lines` and `plot_points`,
  if the x and y limits are are not provided.
}
