% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{extract.figure}
\alias{extract.figure}
\title{Extract features of a figure object.}
\usage{
\method{extract}{figure}(figure_object, what = c("plot", "label", "w", "h", "unit"))
}
\arguments{
\item{figure_object}{A figure object}

\item{what}{The requested feature.
'plot': the ggplot, 'label': the rendering/saving label, 'h': rendering height,
'w': the rendering width, 'unit': the rendering unit.}
}
\value{
The requested feature
}
\description{
Extract features of a figure object.
}
