% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract}
\alias{extract}
\title{Extract object features.}
\usage{
extract(x, ...)
}
\arguments{
\item{x}{An object of interest}

\item{...}{Arguments passed to specific methods}
}
\value{
The requested feature
}
\description{
Extract object features.
}
