% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{height<-}
\alias{height<-}
\title{Set object height.}
\usage{
height(x, lock = FALSE) <- value

height(x, lock = FALSE) <- value
}
\arguments{
\item{x}{A figure object}

\item{value}{New height}

\item{lock}{If TRUE, the aspect ratio is kept constant (scaling)}

\item{...}{Arguments passed to specific methods}
}
\value{
The object with the modified h property

A figure object with the height property modified
}
\description{
Set object height.

Set figure object height.
}
