% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/oop.R
\name{height<-}
\alias{height<-}
\alias{width<-}
\alias{height}
\alias{width}
\alias{resize}
\alias{height.figure}
\alias{width.figure}
\alias{resize.figure}
\title{Get or set set figure object dimensions.}
\usage{
height(x) <- value

width(x) <- value

height(x, ...)

width(x, ...)

resize(x, ...)

\method{height}{figure}(x, ...)

\method{width}{figure}(x, ...)

height(x) <- value

width(x) <- value

\method{resize}{figure}(x, value, what = c("w", "h", "scale"), unit = x$unit, lock = TRUE, ...)
}
\arguments{
\item{x}{a figure class object.}

\item{value}{new height or width or a scaling factor.}

\item{...}{extra arguments passed to methods.}

\item{what}{The requested dimension to be changed or the scaling factor.}

\item{unit}{The dimension unit of the final plot object.}

\item{lock}{logical, if TRUE, the aspect ratio is kept constant (scaling).}
}
\value{
A `figure` object with the height or width property modified
or the requested dimensions.
}
\description{
Get or set dimensions of a `figure` class object.
}
\details{
`height()`, `width()` and `resize()` are S3 generic functions.
}
