% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{height.figure}
\alias{height.figure}
\title{Get figure object's height.}
\usage{
\method{height}{figure}(x, ...)
}
\arguments{
\item{x}{A figure object}
}
\value{
The figure's height
}
\description{
Get figure object's height.
}
