% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{insert.figure}
\alias{insert.figure}
\title{Insert a figure object chunk into an R markdown file}
\usage{
\method{insert}{figure}(
  object,
  file = NULL,
  html = FALSE,
  style_ref = "legend",
  append = TRUE,
  relative_dim = FALSE,
  ...
)
}
\arguments{
\item{object}{a figure object to be referenced (\code{\link{figure}})}

\item{file}{a file to which the chunks are written. If the file exists
already, it will be appended or overwritten.
If NULL, the text is printed in the console and copied into the clipboard.}

\item{html}{logical: should the legend be HTML styled?}

\item{style_ref}{name of the CSS style of the legend text, valid only for
the HTML output.}

\item{append}{logical, should the output file be appended?}

\item{relative_dim}{logical, should the figure dimensions be inserted
as a call to an unit-converting function? If FALSE, fixed numeric dimensions
are inserted into the code chunk.}

\item{...}{extra arguments, currently none.}
}
\value{
returns invisibly the requested R code chunk.
}
\description{
Inserts a reference to the figure object as an R code chunk into
a R markdown file or prints it into the standard output and, concomitantly,
copies into the clipboard.
}
\details{
Only for the figure class instances with
defined 'ref_name' and 'caption' parameters.
To enable the clipboard access, you may need to set the CLIPR_ALLOW
environment variable to TRUE, as described for
\code{\link[clipr]{write_clip}}.
}
