% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_figure}
\alias{is_figure}
\title{Check for a figure object.}
\usage{
is_figure(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, TRUE if the 'figure' class object
}
\description{
Check for a figure object.
}
