% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/oop.R
\name{pickle}
\alias{pickle}
\alias{pickle.figure}
\alias{pickle.mdtable}
\title{Save a figure object or mdtable object on the disc.}
\usage{
pickle(object, ...)

\method{pickle}{figure}(object, path = ".", format = "pdf", ...)

\method{pickle}{mdtable}(object, folder = ".", format = "tsv", delim = "\\t", ...)
}
\arguments{
\item{object}{a figure object.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{ggsave}} (figure)
or \code{\link[readr]{write_delim}} (table).}

\item{path}{The target path. The destination folder will not be created.}

\item{format}{the desired file format.}

\item{folder}{The target folder. The destination folder will not be created.}

\item{delim}{delimiter for column separation, tabulation by default.}
}
\value{
None, called for side effects
}
\description{
Saves a figure object on the disc.
}
\details{
S3 generic function.
}
