% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{Line plot of x and y values}
\usage{
plot_lines(
  x,
  y,
  addPoints = TRUE,
  lineKula = "#004488FF",
  lwd = 4,
  lty = 1,
  borderKula = NULL,
  pointKula = "white",
  cex = 1.5,
  pch = 21,
  xpd = TRUE,
  ...
)
}
\arguments{
\item{x}{The x data to plot.}

\item{y}{The y data to plot.}

\item{addPoints}{LOGICAL: Should points be added on top of the lines (TRUE,
default), or not (FALSE)?}

\item{lineKula}{What colour should the line be?}

\item{lwd}{How thick should the line be?}

\item{lty}{What type (solid, dotted, dashed, etc.) should the line be?}

\item{borderKula}{What colour should the boundary of the marker be?}

\item{pointKula}{What colour should the centre of the marker be?}

\item{cex}{What size should the marker be?}

\item{pch}{What style should the marker be?}

\item{xpd}{Can the markers be plotted off the chart?}
}
\description{
This is the figuR equivalent of `plot(x, y, type = "l")` to
  create a simple line chart.
}
