% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Scatterplot of x and y values}
\usage{
plot_points(
  x,
  y,
  borderKula = "#004488FF",
  pointKula = "white",
  cex = 1.25,
  pch = NULL,
  lwd = 3,
  xpd = TRUE,
  ...
)
}
\arguments{
\item{x}{The x data to plot.}

\item{y}{The y data to plot. y must either be the same length as x, or a
multiple of its length; if the latter, the x data is repeated, and the y
treated as separate data series; 'borderKula' is ignored in such a
situation and each y series is coloured differently (colours will repeat
after 7 y "series").}

\item{borderKula}{What colour should the boundary of the marker be?}

\item{pointKula}{What colour should the centre of the marker be?}

\item{cex}{What size should the marker be?}

\item{pch}{What style should the marker be?}

\item{lwd}{How thick should the marker boundary be?}

\item{xpd}{Can the markers be plotted off the chart?}

\item{...}{Any arguments that can be fed into `pre_plot()`.}
}
\description{
This is the figuR equivalent of `plot(x, y)` to create a
  simple scatterplot.
}
