% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{print.figure}
\alias{print.figure}
\title{Print figure object properties.}
\usage{
\method{print}{figure}(x, ...)
}
\arguments{
\item{x}{A figure object}
}
\value{
The figure's label and dimensions
}
\description{
Print figure object properties.
}
