% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{print.figure}
\alias{print.figure}
\title{Print figure object properties.}
\usage{
\method{print}{figure}(figure_object)
}
\arguments{
\item{figure_object}{A figure object}
}
\value{
The figure's label and dimensions
}
\description{
Print figure object properties.
}
