% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{save_figure}
\alias{save_figure}
\title{Save a plot object on the disc}
\usage{
save_figure(figure_object, path = ".", format = "pdf", ...)
}
\arguments{
\item{figure_object}{A figure object}

\item{path}{The target path. The destination folder will not be created}
}
\value{
None, called for side effects
}
\description{
Save a plot object on the disc
}
