% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{width<-}
\alias{width<-}
\title{Set object width.}
\usage{
width(x, lock = FALSE) <- value

width(x, lock = FALSE) <- value
}
\arguments{
\item{x}{A figure object}

\item{value}{New height}

\item{lock}{If TRUE, the aspect ratio is kept constant (scaling)}

\item{...}{Arguments passed to specific methods}
}
\value{
The object with the modified w property

A figure object with the height property modified
}
\description{
Set object width.

Set figure object width.
}
