% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{width.figure}
\alias{width.figure}
\title{Get figure object's width.}
\usage{
\method{width}{figure}(x, ...)
}
\arguments{
\item{x}{A figure object}
}
\value{
The figure's width
}
\description{
Get figure object's width.
}
